Refactored my channels test to remove distinguishment between your channel and other channels
