+++ Tribler/Test/Core/Modules/RestApi/base_api_test.py
@@ -60,9 +60,7 @@ class AbstractApiTest(AbstractBaseApiTest):
 
     def parse_response(self, response):
         self.assertEqual(response.code, self.expected_response_code)
-        if response.code in (200, 400, 500):
-            return readBody(response)
-        return succeed(None)
+        return readBody(response)
 
     def do_request(self, endpoint, expected_code=200, expected_json=None, request_type='GET', post_data=''):
         self.expected_response_code = expected_code
